+++ src/common/witness/deposit_witness.rs
@@ -6,7 +6,7 @@ use crate::common::{deposit::Deposit, salt::Salt, trees::deposit_tree::DepositMe
 #[serde(rename_all = "camelCase")]
 pub struct DepositWitness {
     pub deposit_salt: Salt,
-    pub deposit_index: usize,
+    pub deposit_index: u32,
     pub deposit: Deposit,
     pub deposit_merkle_proof: DepositMerkleProof,
 }
@@ -16,6 +16,6 @@ pub struct DepositWitness {
 #[serde(rename_all = "camelCase")]
 pub struct DepositCase {
     pub deposit_salt: Salt,
-    pub deposit_index: usize,
+    pub deposit_index: u32,
     pub deposit: Deposit,
 }
